refactor(shared): naming
