+++ src/features/feedback/feedback-instructor/index.tsx
@@ -2,7 +2,7 @@ import Box from "@mui/joy/Box";
 import Typography from "@mui/joy/Typography";
 import Divider from "@mui/joy/Divider";
 
-import BreadcrumbsCustom from "shared/components/breadcrumbsCustom";
+import BreadcrumbsCustom from "shared/components/breadcrumbs-custom";
 import InstructorTickets from "./instructor-tickets";
 
 export default function FeedbackInstructor() {
