+++ src/layouts/dashboard/app.tsx
@@ -2,13 +2,13 @@ import { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
 
 import Root from './wrappers/root';
-import TopNav from './wrappers/topNav';
-import SideNav from './wrappers/sideNav';
+import TopNav from './wrappers/top-nav';
+import SideNav from './wrappers/side-nav';
 import Main from './wrappers/main';
 
 import Header from './components/header';
 import Navigation from './components/navigation';
-import SuspenseCircle from 'shared/components/suspenseCircle';
+import SuspenseCircle from 'shared/components/suspense-circle';
 import RootLayout from 'layouts/root-layout';
 
 export default function DashboardLayout() {
