+++ src/layouts/dashboard/components/header.tsx
@@ -30,12 +30,12 @@ import { authApi } from 'app/base/auth.api';
 
 import { DashboardPath, AuthPath } from 'routes/paths';
 
-import ColorSchemeToggle from 'shared/components/colorToggle';
+import ColorSchemeToggle from 'shared/components/color-toggle';
 
 import LerngLogo from 'assets/svg/logo-no-background.svg';
 
 import style from "../dashboard.module.css";
-import PageSelect from '../utils/pageSelect';
+import PageSelect from '../utils/page-select';
 import TeamNav from "./navigation";
 
 export default function Header() {
