+++ src/layouts/dashboard/components/navigation.tsx
@@ -2,7 +2,7 @@ import List from '@mui/joy/List';
 import ListSubheader from '@mui/joy/ListSubheader';
 import ListItem from '@mui/joy/ListItem';
 
-import RoleBasedList from '../utils/roleBasedList';
+import RoleBasedList from '../utils/role-based-nav';
 
 export default function Navigation() {
   return (
