+++ src/layouts/dashboard/utils/user-nav.tsx
@@ -6,26 +6,32 @@ import ListItemButton from '@mui/joy/ListItemButton';
 import ListItemDecorator from '@mui/joy/ListItemDecorator';
 import ListItemContent from '@mui/joy/ListItemContent';
 
-import DashboardIcon from '@mui/icons-material/Dashboard';
+import LibraryBooksIcon from '@mui/icons-material/LibraryBooks';
+import FeedbackIcon from '@mui/icons-material/Feedback';
 import SignpostIcon from '@mui/icons-material/Signpost';
 
 import { CoursesPath, DashboardPath } from 'routes/paths';
 import style from '../dashboard.module.css';
 
-const InstructorListItems = [
+const UserListItems = [
   {
-    link: DashboardPath.DASHBOARD,
-    name: 'Instructor Panel',
-    icon: <DashboardIcon />
+    link: DashboardPath.ENROLLED,
+    name: 'Enrolled',
+    icon: <LibraryBooksIcon />
   },
   {
     link: CoursesPath.COURSES,
     name: 'Courses',
     icon: <SignpostIcon />
+  },
+  {
+    link: DashboardPath.FEEDBACK,
+    name: 'Feedback',
+    icon: <FeedbackIcon />
   }
 ];
 
-export default function InstructorList() {
+export default function UserNav() {
   return (
     <List
       aria-labelledby="nav-list-browse"
@@ -33,7 +39,7 @@ export default function InstructorList() {
         '& .JoyListItemButton-root': { p: '8px' },
       }}
     >
-      {InstructorListItems.map(item => (
+      {UserListItems.map(item => (
         <ListItem key={item.name}>
           <ListItemButton>
             <ListItemDecorator>
