+++ src/layouts/settings/layout.tsx
@@ -1,6 +1,6 @@
 import Box from '@mui/joy/Box';
 import Typography from '@mui/joy/Typography'
-import BreadcrumbsCustom from 'shared/components/breadcrumbsCustom';
+import BreadcrumbsCustom from 'shared/components/breadcrumbs-custom';
 
 interface LayoutChild {
   children: JSX.Element;
