+++ src/shared/components/alerts/error.tsx
