+++ src/shared/components/color-toggle.tsx
